yield-mvp/agents/ai-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic==1.8.2
motor==2.5.1
prometheus-client==0.12.0
python-dotenv==0.19.0
near-api-py==0.3.1
langchain==0.0.150
transformers==4.11.3
pandas==1.3.3
aiohttp==3.8.1
starlette==0.14.2
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0
python-logging-loki==0.3.1
torch==2.0.0
streamlit==1.22.0
plotly==5.14.1
websockets==10.4
requests==2.28.2
